ref(quick-trace): Refactor quick trace as a single component
Quick trace is current wrapped inside the header component. This change
refactors quick into its own standalone component that can be reused in issues.
