+++ src/sentry/static/sentry/app/views/performance/transactionDetails/styles.tsx
@@ -1,14 +1,9 @@
 import React from 'react';
 import styled from '@emotion/styled';
-import {LocationDescriptor} from 'history';
 
 import {SectionHeading} from 'app/components/charts/styles';
-import MenuItem from 'app/components/menuItem';
 import QuestionTooltip from 'app/components/questionTooltip';
-import Tag, {Background} from 'app/components/tag';
-import Truncate from 'app/components/truncate';
 import space from 'app/styles/space';
-import {Theme} from 'app/utils/theme';
 
 type MetaDataProps = {
   headingText: string;
@@ -51,84 +46,4 @@ const SectionBody = styled('div')`
 export const SectionSubtext = styled('div')<{type?: 'error' | 'default'}>`
   color: ${p => (p.type === 'error' ? p.theme.error : p.theme.subText)};
   font-size: ${p => p.theme.fontSizeMedium};
-`;
-
-const nodeColors = (theme: Theme) => {
-  return {
-    error: {
-      color: theme.white,
-      background: theme.red300,
-      border: theme.red300,
-    },
-    warning: {
-      color: theme.red300,
-      background: theme.white,
-      border: theme.red300,
-    },
-    white: {
-      color: theme.textColor,
-      background: theme.background,
-      border: theme.textColor,
-    },
-    black: {
-      color: theme.background,
-      background: theme.textColor,
-      border: theme.textColor,
-    },
-  };
-};
-
-export const EventNode = styled(Tag)<{pad?: 'left' | 'right'}>`
-  span {
-    color: ${p => nodeColors(p.theme)[p.type || 'white'].color};
-  }
-  & ${/* sc-selector */ Background} {
-    background-color: ${p => nodeColors(p.theme)[p.type || 'white'].background};
-    border: 1px solid ${p => nodeColors(p.theme)[p.type || 'white'].border};
-  }
-`;
-
-export const TraceConnector = styled('div')`
-  width: ${space(1)};
-  border-top: 1px solid ${p => p.theme.textColor};
-`;
-
-export const QuickTraceContainer = styled('div')`
-  display: flex;
-  align-items: center;
-`;
-
-const StyledMenuItem = styled(MenuItem)<{first?: boolean}>`
-  border-top: ${p => (!p.first ? `1px solid ${p.theme.innerBorder}` : null)};
-  width: 350px;
-`;
-
-const MenuItemContent = styled('div')`
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-`;
-
-type DropdownItemProps = {
-  children: React.ReactNode;
-  to?: string | LocationDescriptor;
-  onSelect?: (eventKey: any) => void;
-  first?: boolean;
-};
-
-export function DropdownItem({children, first, onSelect, to}: DropdownItemProps) {
-  return (
-    <StyledMenuItem to={to} onSelect={onSelect} first={first}>
-      <MenuItemContent>{children}</MenuItemContent>
-    </StyledMenuItem>
-  );
-}
-
-export const DropdownItemSubContainer = styled('div')`
-  display: flex;
-  flex-direction: row;
-`;
-
-export const StyledTruncate = styled(Truncate)`
-  margin-left: ${space(1)};
 `;
